Refactored to Onion Architecture
