+++ Chirp.sln
@@ -10,6 +10,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Web", "src\Chirp.Web\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Test", "test\Chirp.Test\Chirp.Test.csproj", "{67FF4EBF-E1A6-4288-A00C-79FF15CBD22C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{F66427B8-AE38-4763-8863-8660DE41D9F1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Infrastructure", "src\Chirp.Infrastructure\Chirp.Infrastructure.csproj", "{2FDE2105-C109-44F4-AD5C-CA2F7FE24961}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,9 +31,19 @@ Global
 		{67FF4EBF-E1A6-4288-A00C-79FF15CBD22C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67FF4EBF-E1A6-4288-A00C-79FF15CBD22C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67FF4EBF-E1A6-4288-A00C-79FF15CBD22C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F66427B8-AE38-4763-8863-8660DE41D9F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F66427B8-AE38-4763-8863-8660DE41D9F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F66427B8-AE38-4763-8863-8660DE41D9F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F66427B8-AE38-4763-8863-8660DE41D9F1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2FDE2105-C109-44F4-AD5C-CA2F7FE24961}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FDE2105-C109-44F4-AD5C-CA2F7FE24961}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FDE2105-C109-44F4-AD5C-CA2F7FE24961}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FDE2105-C109-44F4-AD5C-CA2F7FE24961}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{C63F1052-B74B-4150-A97B-1035225E1015} = {430C3F51-F7D9-4677-9994-BA6EABAA6C38}
 		{67FF4EBF-E1A6-4288-A00C-79FF15CBD22C} = {7420877A-8C4F-48A4-87A6-9F70A959BF75}
+		{F66427B8-AE38-4763-8863-8660DE41D9F1} = {430C3F51-F7D9-4677-9994-BA6EABAA6C38}
+		{2FDE2105-C109-44F4-AD5C-CA2F7FE24961} = {430C3F51-F7D9-4677-9994-BA6EABAA6C38}
 	EndGlobalSection
 EndGlobal
