+++ src/Chirp.Infrastructure/Chirp.Infrastructure.csproj
@@ -1,0 +1,24 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <Content Include="Migrations\20241002120734_InitialDBSchema.cs" />
+      <Content Include="Migrations\20241002120734_InitialDBSchema.Designer.cs" />
+      <Content Include="Migrations\CheepDBContextModelSnapshot.cs" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.0-rc.2.24474.1" />
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.0-rc.2.24474.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Chirp.Core\Chirp.Core.csproj" />
+    </ItemGroup>
+
+</Project>
