+++ src/Chirp.Web/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=Cheep.db"
+    "DefaultConnection": "Data Source=..\\Chirp.Infrastructure\\Data\\Cheep.db"
   },
   "AllowedHosts": "*"
 }
