+++ test/Chirp.Test/IntegrationTests.cs
@@ -3,13 +3,15 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
+using Chirp.Core;
+using Chirp.Infrastructure.Repositories;
 using FluentAssertions;
-using Chirp.Razor;
+using Chirp.Web;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Data.Sqlite;
 
-namespace Chirp.Razor.test;
+namespace Chirp.Test;
 
 public class IntegrationTests : IClassFixture<WebApplicationFactory<Program>>, IDisposable
 {
