+++ test/Chirp.Test/UnitTests.cs
@@ -1,3 +1,5 @@
+using Chirp.Core;
+using Chirp.Infrastructure.Repositories;
 using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
